ASoC: msm: qdsp6v2: Force V3 functions for MSM8996
The drivers are sometimes calling the V4 functions directly for
some monkey reason...
This makes the DSP to go completely nuts since the firmware
is not updated to support the V4 functions.

To address this, even if - okay, it's ugly - add ifdefs to force
calling the corresponding V3 functions: this makes sure that we
won't get breakage from random CAF tags upgrades and such.

Signed-off-by: Subhajeet Muhuri <kenny3fcb@gmail.com>
Signed-off-by: Yaroslav Furman <yaro330@gmail.com>